[codegen] Split up parsing module
- move parsing of attributes & fields out to a new parsing::attrs
  and parsing::fields modules
- moves 'NeededWhen' into fields.rs, where it is mainly used.
- moves a few other stray items into more sensible locations
